Remove all address index based code into address and remove calls from EquivAddress
