+++ src/blocksci/address/address.cpp
@@ -21,6 +21,11 @@
 #include "index/address_index.hpp"
 #include "index/hash_index.hpp"
 
+#include <range/v3/view/transform.hpp>
+#include <range/v3/view/filter.hpp>
+#include <range/v3/action/unique.hpp>
+#include <range/v3/action/sort.hpp>
+
 #include <unordered_set>
 #include <iostream>
 
@@ -70,13 +75,13 @@ namespace blocksci {
     uint64_t Address::calculateBalance(BlockHeight height, bool typeEquivalent, bool nestedEquivalent) const {
         uint64_t value = 0;
         if (height == 0) {
-            for (auto &output : access->getOutputs(*this, typeEquivalent, nestedEquivalent)) {
+            for (auto &output : getOutputs(typeEquivalent, nestedEquivalent)) {
                 if (!output.isSpent()) {
                     value += output.getValue();
                 }
             }
         } else {
-            for (auto &output : access->getOutputs(*this, typeEquivalent, nestedEquivalent)) {
+            for (auto &output : getOutputs(typeEquivalent, nestedEquivalent)) {
                 if (output.blockHeight <= height && (!output.isSpent() || output.getSpendingTx()->blockHeight > height)) {
                     value += output.getValue();
                 }
@@ -85,25 +90,7 @@ namespace blocksci {
         return value;
     }
     
-    std::vector<Output> Address::getOutputs(bool typeEquivalent, bool nestedEquivalent) const {
-        return access->getOutputs(*this, typeEquivalent, nestedEquivalent);
-    }
-    
-    std::vector<Input> Address::getInputs(bool typeEquivalent, bool nestedEquivalent) const {
-        return access->getInputs(*this, typeEquivalent, nestedEquivalent);
-    }
-    
-    std::vector<Transaction> Address::getTransactions(bool typeEquivalent, bool nestedEquivalent) const {
-        return access->getTransactions(*this, typeEquivalent, nestedEquivalent);
-    }
-    
-    std::vector<Transaction> Address::getOutputTransactions(bool typeEquivalent, bool nestedEquivalent) const {
-        return access->getOutputTransactions(*this, typeEquivalent, nestedEquivalent);
-    }
-    
-    std::vector<Transaction> Address::getInputTransactions(bool typeEquivalent, bool nestedEquivalent) const {
-        return access->getInputTransactions(*this, typeEquivalent, nestedEquivalent);
-    }
+
     
     ranges::optional<Address> getAddressFromString(const std::string &addressString, const DataAccess &access) {
         if (addressString.compare(0, access.config.segwitPrefix.size(), access.config.segwitPrefix) == 0) {
@@ -193,6 +180,68 @@ namespace blocksci {
 
     std::string Address::fullType() const {
         return fullTypeImp(*this, *access);
+    }
+    
+    std::vector<Output> Address::getOutputs(bool typeEquivalent, bool nestedEquivalent) const {
+        auto dataAccess = access;
+        auto pointers = access->addressIndex->getOutputPointers(*this, typeEquivalent, nestedEquivalent);
+        return pointers
+        | ranges::view::transform([dataAccess](const OutputPointer &pointer) { return Output(pointer, *dataAccess); })
+        | ranges::to_vector;
+    }
+    
+    std::vector<Input> Address::getInputs(bool typeEquivalent, bool nestedEquivalent) const {
+        auto pointers = access->addressIndex->getOutputPointers(*this, typeEquivalent, nestedEquivalent);
+        std::unordered_set<InputPointer> allPointers;
+        allPointers.reserve(pointers.size());
+        for (auto &pointer : pointers) {
+            auto inputTx = Output(pointer, *access).getSpendingTx();
+            if(inputTx) {
+                auto inputPointers = inputTx->getInputPointers(pointer);
+                for (auto &inputPointer : inputPointers) {
+                    allPointers.insert(inputPointer);
+                }
+            }
+        }
+        auto dataAccess = access;
+        return allPointers
+        | ranges::view::transform([dataAccess](const InputPointer &pointer) { return Input(pointer, *dataAccess); })
+        | ranges::to_vector;
+    }
+    
+    std::vector<Transaction> Address::getTransactions(bool typeEquivalent, bool nestedEquivalent) const {
+        auto pointers = access->addressIndex->getOutputPointers(*this, typeEquivalent, nestedEquivalent);
+        std::unordered_set<blocksci::Transaction> txes;
+        txes.reserve(pointers.size() * 2);
+        for (auto &pointer : pointers) {
+            txes.insert(Transaction(pointer.txNum, *access));
+            auto inputTx = Output(pointer, *access).getSpendingTx();
+            if (inputTx) {
+                txes.insert(*inputTx);
+            }
+        }
+        return {txes.begin(), txes.end()};
+    }
+    
+    std::vector<Transaction> Address::getOutputTransactions(bool typeEquivalent, bool nestedEquivalent) const {
+        auto pointers = access->addressIndex->getOutputPointers(*this, typeEquivalent, nestedEquivalent);
+        auto txNums = pointers | ranges::view::transform([](const OutputPointer &pointer) { return pointer.txNum; }) | ranges::to_vector;
+        txNums |= ranges::action::sort | ranges::action::unique;
+        auto dataAccess = access;
+        return txNums | ranges::view::transform([dataAccess](uint32_t txNum) { return Transaction(txNum, *dataAccess); }) | ranges::to_vector;
+    }
+
+    auto flatMap() {
+        return ranges::view::filter([](const auto &optional) { return static_cast<bool>(optional); })
+        | ranges::view::transform([](const auto &optional) { return *optional; });
+    }
+    
+    std::vector<Transaction> Address::getInputTransactions(bool typeEquivalent, bool nestedEquivalent) const {
+        auto pointers = access->addressIndex->getOutputPointers(*this, typeEquivalent, nestedEquivalent);
+        auto dataAccess = access;
+        auto txes = pointers | ranges::view::transform([dataAccess](const OutputPointer &pointer) { return Output(pointer, *dataAccess).getSpendingTx(); }) | flatMap() | ranges::to_vector;
+        txes |= ranges::action::sort | ranges::action::unique;
+        return txes;
     }
 }
