+++ src/blocksci/address/equiv_address.cpp
@@ -40,44 +40,6 @@ namespace blocksci {
             return ss.str();
         }
     }
-    
-    uint64_t EquivAddress::calculateBalance(BlockHeight height) const {
-        uint64_t value = 0;
-        if (height == 0) {
-            for (auto &output : access->getOutputs(*this)) {
-                if (!output.isSpent()) {
-                    value += output.getValue();
-                }
-            }
-        } else {
-            for (auto &output : access->getOutputs(*this)) {
-                if (output.blockHeight <= height && (!output.isSpent() || output.getSpendingTx()->blockHeight > height)) {
-                    value += output.getValue();
-                }
-            }
-        }
-        return value;
-    }
-    
-    std::vector<Output> EquivAddress::getOutputs() const {
-        return access->getOutputs(*this);
-    }
-    
-    std::vector<Input> EquivAddress::getInputs() const {
-        return access->getInputs(*this);
-    }
-    
-    std::vector<Transaction> EquivAddress::getTransactions() const {
-        return access->getTransactions(*this);
-    }
-    
-    std::vector<Transaction> EquivAddress::getOutputTransactions() const {
-        return access->getOutputTransactions(*this);
-    }
-    
-    std::vector<Transaction> EquivAddress::getInputTransactions() const {
-        return access->getInputTransactions(*this);
-    }
 }
 
 std::ostream &operator<<(std::ostream &os, const blocksci::EquivAddress &address) {
